# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Api",
    "version": "0.0.1"
  },
  "paths": {
...
    "/align/scan-source": {
      "post": {
        "tags": ["align"],
        "operationId": "align.scanSource",
        "parameters": [],
        "security": [],
        "responses": {
          "200": {
            "description": "WorkspaceScan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceScan"
                }
              }
            }
          },
          "400": {
            "description": "The request did not match the expected schema",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/HttpApiDecodeError"
                    },
                    {
                      "$ref": "#/components/schemas/RequestError"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanSourceRequest"
              }
            }
          },
          "required": true
        }
      }
    },
    "/align/load-frame": {
      "post": {
        "tags": ["align"],
        "operationId": "align.loadFrame",
        "parameters": [],
        "security": [],
        "responses": {
          "200": {
            "description": "FramePayload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FramePayload"
                }
              }
            }
          },
          "400": {
            "description": "The request did not match the expected schema",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/HttpApiDecodeError"
                    },
                    {
                      "$ref": "#/components/schemas/RequestError"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoadFrameRequest"
              }
            }
          },
          "required": true
        }
      }
    },
    "/align/save-bbox": {
      "post": {
        "tags": ["align"],
        "operationId": "align.saveBbox",
        "parameters": [],
        "security": [],
        "responses": {
          "200": {
            "description": "SaveBboxResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaveBboxResponse"
                }
              }
            }
          },
          "400": {
            "description": "The request did not match the expected schema",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/HttpApiDecodeError"
                    },
                    {
                      "$ref": "#/components/schemas/RequestError"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveBboxRequest"
              }
            }
          },
          "required": true
        }
      }
    },
    "/align/align-state": {
      "get": {
        "tags": ["align"],
        "operationId": "align.loadAlignState",
        "parameters": [],
        "security": [],
        "responses": {
          "200": {
            "description": "SavedAlignState",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/SavedAlignState"
                    },
                    {
                      "type": "null"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "The request did not match the expected schema",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/HttpApiDecodeError"
                    },
                    {
                      "$ref": "#/components/schemas/RequestError"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          }
        }
      }
    },
    "/align/output-paths": {
      "get": {
        "tags": ["align"],
        "operationId": "align.outputPaths",
        "parameters": [],
        "security": [],
        "responses": {
          "200": {
            "description": "AlignOutputPaths",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlignOutputPaths"
                }
              }
            }
          },
          "400": {
            "description": "The request did not match the expected schema",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/HttpApiDecodeError"
                    },
                    {
                      "$ref": "#/components/schemas/RequestError"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          }
        }
      }
    },
    "/align/saved-bbox-positions": {
      "get": {
        "tags": ["align"],
        "operationId": "align.listSavedBboxPositions",
        "parameters": [],
        "security": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "integer",
                    "format": "uint32",
                    "minimum": 0
                  }
                }
              }
            }
          },
          "400": {
            "description": "The request did not match the expected schema",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/HttpApiDecodeError"
                    },
                    {
                      "$ref": "#/components/schemas/RequestError"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          }
        }
      }
    },
    "/align/roi-pos-exists": {
      "get": {
        "tags": ["align"],
        "operationId": "align.roiPosExists",
        "parameters": [],
        "security": [],
        "responses": {
          "200": {
            "description": "RoiPosExistsResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoiPosExistsResponse"
                }
              }
            }
          },
          "400": {
            "description": "The request did not match the expected schema",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/HttpApiDecodeError"
                    },
                    {
                      "$ref": "#/components/schemas/RequestError"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          }
        }
      }
    },
    "/align/crop-roi": {
      "post": {
        "tags": ["align"],
        "operationId": "align.cropRoi",
        "parameters": [],
        "security": [],
        "responses": {
          "200": {
            "description": "CropRoiResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CropRoiResponse"
                }
              }
            }
          },
          "400": {
            "description": "The request did not match the expected schema",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/HttpApiDecodeError"
                    },
                    {
                      "$ref": "#/components/schemas/RequestError"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CropRoiRequest"
              }
            }
          },
          "required": true
        }
      }
    },
    "/align/cancel-crop-roi": {
      "post": {
        "tags": ["align"],
        "operationId": "align.cancelCropRoi",
        "parameters": [],
        "security": [],
        "responses": {
          "200": {
            "description": "CropRoiProgress",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CropRoiProgress"
                }
              }
            }
          },
          "400": {
            "description": "The request did not match the expected schema",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/HttpApiDecodeError"
                    },
                    {
                      "$ref": "#/components/schemas/RequestError"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelCropRoiRequest"
              }
            }
          },
          "required": true
        }
      }
    },
    "/align/crop-roi-progress": {
      "get": {
        "tags": ["align"],
        "operationId": "align.cropRoiProgress",
        "parameters": [],
        "security": [],
        "responses": {
          "200": {
            "description": "CropRoiProgress",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CropRoiProgress"
                }
              }
            }
          },
          "400": {
            "description": "The request did not match the expected schema",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/HttpApiDecodeError"
                    },
                    {
                      "$ref": "#/components/schemas/RequestError"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          }
        }
      }
    },
    "/align/crop-latest": {
      "get": {
        "tags": ["align"],
        "operationId": "align.getLatestCropProgress",
        "parameters": [],
        "security": [],
        "responses": {
          "200": {
            "description": "CropRoiProgress",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/CropRoiProgress"
                    },
                    {
                      "type": "null"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "The request did not match the expected schema",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/HttpApiDecodeError"
                    },
                    {
                      "$ref": "#/components/schemas/RequestError"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HttpApiDecodeError": {
        "type": "object",
        "required": ["issues", "message", "_tag"],
        "properties": {
          "issues": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Issue"
            }
          },
          "message": {
            "type": "string"
          },
          "_tag": {
            "type": "string",
            "enum": ["HttpApiDecodeError"]
          }
        },
        "additionalProperties": false,
        "description": "The request did not match the expected schema"
      },
      "Issue": {
        "type": "object",
        "required": ["_tag", "path", "message"],
        "properties": {
          "_tag": {
            "type": "string",
            "enum": [
              "Pointer",
              "Unexpected",
              "Missing",
              "Composite",
              "Refinement",
              "Transformation",
              "Type",
              "Forbidden"
            ],
            "description": "The tag identifying the type of parse issue"
          },
          "path": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PropertyKey"
            },
            "description": "The path to the property where the issue occurred"
          },
          "message": {
            "type": "string",
            "description": "A descriptive message explaining the issue"
          }
        },
        "additionalProperties": false,
        "description": "Represents an error encountered while parsing a value to match the schema"
      },
...
      "RequestError": {
        "type": "object",
        "required": ["message", "_tag"],
        "properties": {
          "message": {
            "type": "string"
          },
          "_tag": {
            "type": "string",
            "enum": ["RequestError"]
          }
        },
        "additionalProperties": false
      },
      "Unauthorized": {
        "type": "object",
        "required": ["message", "_tag"],
        "properties": {
          "message": {
            "type": "string"
          },
          "_tag": {
            "type": "string",
            "enum": ["Unauthorized"]
          }
        },
        "additionalProperties": false
      },
...
      "ScanSourceRequest": {
        "type": "object",
        "required": ["source"],
        "properties": {
          "source": {
            "$ref": "#/components/schemas/AlignerSource"
          }
        },
        "additionalProperties": false
      },
      "AlignerSource": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/FolderSource"
          },
          {
            "type": "object",
            "required": ["kind", "path"],
            "properties": {
              "kind": {
                "type": "string",
                "enum": ["nd2"]
              },
              "path": {
                "type": "string"
              }
            },
            "additionalProperties": false
          },
          {
            "type": "object",
            "required": ["kind", "path"],
            "properties": {
              "kind": {
                "type": "string",
                "enum": ["czi"]
              },
              "path": {
                "type": "string"
              }
            },
            "additionalProperties": false
          }
        ]
      },
...
      "WorkspaceScan": {
        "type": "object",
        "required": ["positions", "channels", "times", "zSlices"],
        "properties": {
          "positions": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "uint32",
              "minimum": 0
            }
          },
          "channels": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "uint32",
              "minimum": 0
            }
          },
          "times": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "uint32",
              "minimum": 0
            }
          },
          "zSlices": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "uint32",
              "minimum": 0
            }
          },
          "positionLabels": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "channelLabels": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "timeLabels": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "zSliceLabels": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "additionalProperties": false
      },
      "LoadFrameRequest": {
        "type": "object",
        "required": ["source", "request", "contrast"],
        "properties": {
          "source": {
            "$ref": "#/components/schemas/AlignerSource"
          },
          "request": {
            "$ref": "#/components/schemas/FrameRequest"
          },
          "contrast": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContrastWindow"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false
      },
...
      "FramePayload": {
        "type": "object",
        "required": [
          "width",
          "height",
          "dataBase64",
          "pixelType",
          "contrastDomain",
          "suggestedContrast",
          "appliedContrast"
        ],
        "properties": {
          "width": {
            "type": "integer",
            "format": "uint32",
            "minimum": 0
          },
          "height": {
            "type": "integer",
            "format": "uint32",
            "minimum": 0
          },
          "dataBase64": {
            "type": "string"
          },
          "pixelType": {
            "$ref": "#/components/schemas/PixelType"
          },
          "contrastDomain": {
            "$ref": "#/components/schemas/ContrastWindow"
          },
          "suggestedContrast": {
            "$ref": "#/components/schemas/ContrastWindow"
          },
          "appliedContrast": {
            "$ref": "#/components/schemas/ContrastWindow"
          }
        },
        "additionalProperties": false
      },
      "PixelType": {
        "type": "string",
        "enum": ["uint8", "uint8clamped", "int8", "uint16", "int16", "uint32", "int32"]
      },
      "SaveBboxRequest": {
        "type": "object",
        "required": ["workspacePath", "pos", "csv", "alignState"],
        "properties": {
          "workspacePath": {
            "type": "string"
          },
          "pos": {
            "type": "integer",
            "format": "uint32",
            "minimum": 0
          },
          "csv": {
            "type": "string"
          },
          "alignState": {
            "$ref": "#/components/schemas/SavedAlignState"
          }
        },
        "additionalProperties": false
      },
      "SavedAlignState": {
        "type": "object",
        "required": ["grid", "excludedCells"],
        "properties": {
          "grid": {
            "$ref": "#/components/schemas/AlignGridState"
          },
          "excludedCells": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AlignGridCellCoord"
            }
          }
        },
        "additionalProperties": false
      },
...
      "SaveBboxResponse": {
        "type": "object",
        "required": ["ok", "error"],
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false
      },
...
      "AlignOutputPaths": {
        "type": "object",
        "required": ["bbox", "align", "roi"],
        "properties": {
          "bbox": {
            "type": "string"
          },
          "align": {
            "type": "string"
          },
          "roi": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "RoiPosExistsResponse": {
        "type": "object",
        "required": ["exists"],
        "properties": {
          "exists": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "CropRoiRequest": {
        "type": "object",
        "required": ["requestId", "workspacePath", "source", "positions", "overwrite"],
        "properties": {
          "requestId": {
            "type": "string"
          },
          "workspacePath": {
            "type": "string"
          },
          "source": {
            "$ref": "#/components/schemas/AlignerSource"
          },
          "positions": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "uint32",
              "minimum": 0
            }
          },
          "overwrite": {
            "type": "boolean"
          },
          "outputFormat": {
            "$ref": "#/components/schemas/CropOutputFormat"
          }
        },
        "additionalProperties": false
      },
      "CropOutputFormat": {
        "type": "string",
        "enum": ["tiff"]
      },
      "CropRoiResponse": {
        "type": "object",
        "required": ["requestId", "status"],
        "properties": {
          "requestId": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/CropRoiStatus"
          }
        },
        "additionalProperties": false
      },
      "CropRoiStatus": {
        "type": "string",
        "enum": ["queued", "running", "completed", "cancelled", "error"]
      },
      "CancelCropRoiRequest": {
        "type": "object",
        "required": ["requestId"],
        "properties": {
          "requestId": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "CropRoiProgress": {
        "type": "object",
        "required": [
          "requestId",
          "status",
          "position",
          "completedPositions",
          "totalPositions",
          "completedRois",
          "totalRois",
          "message"
        ],
        "properties": {
          "requestId": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/CropRoiStatus"
          },
          "position": {
            "anyOf": [
              {
                "type": "integer",
                "format": "uint32",
                "minimum": 0
              },
              {
                "type": "null"
              }
            ]
          },
          "completedPositions": {
            "type": "integer",
            "format": "uint32",
            "minimum": 0
          },
          "totalPositions": {
            "type": "integer",
            "format": "uint32",
            "minimum": 0
          },
          "completedRois": {
            "type": "integer",
            "format": "uint32",
            "minimum": 0
          },
          "totalRois": {
            "type": "integer",
            "format": "uint32",
            "minimum": 0
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "skippedPositions": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "uint32",
              "minimum": 0
            }
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {}
  },
  "security": [],
...
}